fix: use HTTPS for argocd-server HTTP route
Update the Argo CD HTTP route to use HTTPS instead of HTTP

This change ensures secure communication with the Argo CD server
by using HTTPS for the internal gateway route.  Previously, the
route used HTTP, which is less secure.  This improves the overall
security posture of the Argo CD deployment.